requirements.txt
amadeus==12.0.0
python-telegram-bot==21.4
python-dotenv
pandas
sqlite-utils
